build out game functionality
deleted fsm. when trying to integrate it with the server, I found that it was hard to use and not useful... implemented Game struct that just maintains the state of a game, nothing more. it has some checks in place to ensure the integrity of the data inside, but that's it. GameRunner is a new idea, it runs the rounds of the game using channels to communicate the important points of the game. will try to integrate it with the sever next...